Enable BLTouch + change IO5 function
IO5 was formerly used for SDSS (SPI chip select). It has been repurposed to pull down the nOE signal, which is used by a bunch of IO chips to disable their output until the ESP32 has started. IO5 is pulled up on boot. It also does some PWM after boot (why?) but this doesn't seem to be an issue for our purpose